Allow specifying how the file paths should be printed.
print-files now taken an optional argument to specify if the file path that is printed should be relative or absolute, for relative you can also specify the base directory the printed path
should be relative to.
print-files without arguments keeps the old behavior of just printing what is specified as the output directory plus the file path.

starting