Build glibd library from autogenerated bindings
This is messing with Meson quite a bit, but ensures that all tools and
libraries are built with the correct flags.
We need to create a temporary girtod binary at configure time, to have
it build the D sources to generate the new library from. This ensures
Meson knows about all actual D sources in advance before the final ninja
run.