Not everything with an elementType is an array.
The blacklist of Container types works for GLib but whould fail for other libraries.
So only considder types defined with the arrat tag in the gir file as an array.
With the exception of the GLib array types which are defined with an array tag but shouldn't be wrapped a s an array.